requirements.txt
appdirs==1.4.3
coreapi==2.3.0
coreschema==0.0.4
dj-stripe==0.8.0
Django==1.9.12
django-allauth==0.31.0
django-braces==1.11.0
django-model-utils==2.6.1
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
django-webpack-loader==0.4.1
djangorestframework==3.6.2
et-xmlfile==1.0.1
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.5
jsonfield==2.0.1
Markdown==2.6.8
MarkupSafe==1.0
oauthlib==2.0.2
openapi-codec==1.3.1
openpyxl==2.4.5
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-openid==2.2.5
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
stripe==1.51.0
uritemplate==3.0.0
